chore: fix problematic astronomia export
node uses package.json exports bundlers may use file access for common
js modules. In the astronomia package for both scenarios a working
approach was found. In this module the require needs to be replaced with
a direct file access to make this happen.
Maybe on the long run this postbuild process can be removed.